First pass at keyboard

+++ MonoGameGum/Forms/FormsUtilities.cs
@@ -306,7 +306,7 @@ public class FormsUtilities
 #if XNALIKE
     public static void Update(Game game, GameTime gameTime, GraphicalUiElement rootGue)
 #else
-    public static void Update(float gameTime, GraphicalUiElement rootGue)
+    public static void Update(double gameTime, GraphicalUiElement rootGue)
 #endif
     {
         innerRootList.Clear();
@@ -324,7 +324,7 @@ public class FormsUtilities
 #if XNALIKE
     public static void Update(Game game, GameTime gameTime, IEnumerable<GraphicalUiElement> roots)
 #else
-    public static void Update(float gameTime, IEnumerable<GraphicalUiElement> roots)
+    public static void Update(double gameTime, IEnumerable<GraphicalUiElement> roots)
 #endif
     {
 #if XNALIKE
